pyproject.toml
[tool.poetry]
name = "emotiondetection-fromtext"
version = "0.1.0"
description = ""
authors = ["Naoya_Nakagawa_Sivax <n-nakagawa_d1@sivax.co.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
janome = "^0.5.0"
pykakasi = "^2.3.0"
pillow = "^10.4.0"
transforms = "^0.2.1"
transformers = "^4.46.3"
matplotlib = "^3.9.2"
pandas = "^2.2.3"
unidecode = "^1.3.8"
mecab = "^0.996.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
